Adds reprotest to confirm reproducibility
Generates a "tinyconfig" so that builds are quick. (There's a `make
tinyconfig` option in the upstream kernel.) Tweaks the tooling somewhat
to accommodate. Adds new reprotest target and runs it in CI. There are a
few tests we skip for reproducibility, but the suite that's passing is
impressive.

In order to get full reproducibility on a clean tarball, rather than
reusing the tarball emitted from the build, I had to patch the debian
changelog logic to support SOURCE_DATE_EPOCH. The script already had
support for custom patches, so it was easy to add.