CI: Reproducibility checks for SD kernels
The CI builds using reprotest for SD aren't the same build process,
mostly because of the lack of grsecurity patches. If we add the creds to
CI and disable building from forked PRs (so creds aren't
world-readable), that'd be a better test.

We're still using the same verbose config, however, which is a good
start, and also means the build takes substantially longer than the
vanilla target: 2.5h vs. 0.5h, or thereabouts.

The long-running build times out in CI when creating the dbg package.
Even with a no_output_timeout of 1h, the reprotest run fails because the
entire build task exceeds 5h. Disabling for now, but leaving in the CCI
config in case a slimmer config makes it viable again.