Hack: Build metapackage with explicit dependencies on old versions
As described in
<https://github.com/freedomofpress/securedrop/issues/6762>, we are going
to have unattended-upgrades automatically remove old packages that are
no longer needed to remove old kernel packages. But since we want to
keep 2 kernels, we need to explicitly depend on the current and previous
one, which is the point of this patch.

This is a one-off hack, in the future our kernel versions will be
recognized by apt/unattended-upgrades as kernels, so apt's "keep 2"
protection will automatically apply.