Explicitly depend on linux-image-5.15.89-grsec-securedrop (temporarily)
Since we want to keep 2 kernels, we need an explicit dependency on
the previous 5.15.89, otherwise it'll be removed by unattended-upgrades.

This is a temporary requirement that should be reverted after this
kernel release; now that the version format has been corrected
apt will keep 2 kernels by itself.

See <https://github.com/freedomofpress/securedrop/issues/6762#issuecomment-1520569189>
for more details (this is step 3).